RW-433 moved ParticipantConditionDbInfo to new package.
